Move the biotools login to manage account
