+++ src/store/access.js
@@ -74,6 +74,16 @@ export const useTokenStore = defineStore({
       await this.syncTokens();
     },
 
+    async saveUsernamePassword(key, userName, password) {
+      let tokenObject = {
+        name: userName,
+        type: "password",
+      };
+      tokenObject.token = await encrypt(password);
+      this.accessTokens[key] = tokenObject;
+      await this.syncTokens();
+    },
+
     async writeDatasetsToFile() {
       fs.ensureFileSync(TOKEN_STORE_PATH);
       fs.writeJsonSync(TOKEN_STORE_PATH, this.accessTokens);
@@ -87,6 +97,43 @@ export const useTokenStore = defineStore({
       if (key in this.accessTokens) {
         const tokenObject = Object.assign({}, this.accessTokens[key]);
         tokenObject.token = await decrypt(this.accessTokens[key].token);
+
+        if (key === "biotools") {
+          const serverURL = this.getServerUrl;
+
+          const url =
+            serverURL !== undefined
+              ? `${serverURL}/biotools/login`
+              : `http://127.0.0.1:7632/biotools/login`;
+          console.log(url);
+
+          const response = await axios
+            .post(url, {
+              username: this.accessTokens[key].name,
+              password: tokenObject.token,
+            })
+            .then(async (response) => {
+              if ("general_errors" in response.data) {
+                return "NO_TOKEN_FOUND";
+              }
+
+              if ("key" in response.data) {
+                const token = response.data.key;
+                return token;
+              }
+            })
+            .catch((error) => {
+              console.error(error);
+              return "NO_TOKEN_FOUND";
+            });
+
+          if (response === "NO_TOKEN_FOUND") {
+            return "NO_TOKEN_FOUND";
+          } else {
+            tokenObject.token = response;
+          }
+        }
+
         return tokenObject;
       } else {
         return "NO_TOKEN_FOUND";
@@ -140,11 +187,11 @@ export const useTokenStore = defineStore({
 
     async verifyBioToolsToken(token) {
       const serverURL = this.getServerUrl;
-      console.log(serverURL);
+
       const url =
         serverURL !== undefined
           ? `${serverURL}/biotools/user?token=${token}`
-          : `http://127.0.0.1:$7632/biotools/user?token=${token}`;
+          : `http://127.0.0.1:7632/biotools/user?token=${token}`;
 
       const config = {
         method: "get",
@@ -156,7 +203,8 @@ export const useTokenStore = defineStore({
           return { data: response.data, status: response.status };
         })
         .catch((error) => {
-          return { data: error.response, status: error.response.status };
+          console.error(error);
+          return { data: error.response };
         });
 
       if (response.status === 200) {
@@ -219,7 +267,7 @@ export const useTokenStore = defineStore({
           return { scope: response.headers["x-oauth-scopes"] };
         })
         .catch((error) => {
-          console.log(error);
+          console.error(error);
           return "Error Found";
         });
     },
